requirements.txt
# Google Cloud Dependencies
google-cloud-dialogflow-cx==1.30.0
google-cloud-aiplatform==1.38.0
google-generativeai==0.3.1
google-cloud-pubsub==2.18.4
google-cloud-logging==3.8.0
google-cloud-secret-manager==2.16.4

# Web Framework
flask==3.0.0
gunicorn==21.2.0

# HTTP & API
requests==2.31.0
pydantic==2.5.0

# Async Processing
celery==5.3.4
redis==5.0.1

# Data Processing
python-dateutil==2.8.2

# Security & Auth
PyJWT==2.8.0
cryptography==41.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Code Quality
pylint==3.0.3
black==23.12.1
